pyproject.toml
[tool.poetry]
name = "ml-carbucks"
version = "0.1.0"
description = ""
authors = [
    "Damian Trafiale <damian.trafiale@gmail.com>",
    "Ricardo Fernandes <ricardo_Fernandes21@hotmail.com>",
    "Nural Hassan <nuralh777@gmail.com>"
]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
ultralytics = "^8.3.203"
torch = "^2.8.0"
torchvision = "^0.23.0"
optuna = "^4.5.0"
optuna-dashboard = "^0.19.0"
torchaudio = "^2.8.0"
transformers = "^4.56.2"
datasets = "^4.1.1"
evaluate = "^0.4.6"
pillow = "^11.3.0"
matplotlib = "^3.10.6"
timm = "^1.0.20"
accelerate = "^1.10.1"
pycocotools = "^2.0.10"
cloudpickle = "^3.1.1"
imagehash = "^4.3.2"


[tool.poetry.group.dev.dependencies]
jupyter = "^1.1.1"
ipykernel = "^6.30.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
